pyproject.toml
[tool.poetry]
name= 'link-shortener-poetry'
version = '1.0'
description = 'Package manager for link shortener'
authors = ['Vojtech Janousek']

[tool.poetry.dependencies]
python = '^3.7'
sanic = '^19.9'
mysql-connector = '^2.2.9'

[tool.poetry.dev-dependencies]
flake8 = '^3.7'

[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'
